+++ src/Products/MeetingLalouviere/adapters.py
@@ -99,11 +99,16 @@ LLO_WAITING_ADVICES_FROM_STATES = {
 }
 adaptations.WAITING_ADVICES_FROM_STATES.update(LLO_WAITING_ADVICES_FROM_STATES)
 
+
 CUSTOM_RETURN_TO_PROPOSING_GROUP_MAPPINGS = {
-    "backTo_itempublished_from_returned_to_proposing_group": [
-        "published",
-        "decided"
-    ]
+    "backTo_itemfrozen_from_returned_to_proposing_group": {
+        "*": ["frozen", "decided", "decisions_published", ],
+        "meeting-config-council": ["frozen", "decisions_published", ]
+    },
+    "backTo_itempublished_from_returned_to_proposing_group": {
+        "*": ["published",],
+        "meeting-config-council": ["published", "decided"]
+    },
 }
 adaptations.RETURN_TO_PROPOSING_GROUP_MAPPINGS.update(CUSTOM_RETURN_TO_PROPOSING_GROUP_MAPPINGS)